+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [2023-05-18]
+- Add tls connection options for ldap connection
 
 ## [2023-05-17]
 - First Commit